Move GCS region path checks into iris.marin_fs
