+++ lib/marin/src/marin/training/training.py
@@ -26,8 +26,7 @@ from levanter.main.train_dpo import TrainDpoConfig
 from levanter.main.train_lm import TrainLmConfig
 from mergedeep import mergedeep
 
-from iris.marin_fs import check_path_in_region, marin_region, marin_temp_bucket
-from marin.utilities.gcs_utils import check_gcs_paths_same_region
+from iris.marin_fs import check_gcs_paths_same_region, check_path_in_region, marin_region, marin_temp_bucket
 
 logger = logging.getLogger(__name__)
